### 2017-08-11 #359 @whedon
[REVIEW]: iheatmapr: Interactive complex heatmaps in R
**Submitting author:** @AliciaSchep (<a href="http://orcid.org/0000-0002-3915-0618">Alicia Schep</a>)
**Repository:** <a href="https://github.com/ropensci/iheatmapr" target ="_blank">https://github.com/ropensci/iheatmapr</a>
**Version:** v0.3.2
