### 2017-08-13 #359 @maelle
This submission has been accepted to rOpenSci. The review thread can be found at https://github.com/ropensci/onboarding/issues/107
