### 2017-08-14 #359 @karthik
I notified @rlbarter about the mix up.

The review thread from ropensci is here: https://github.com/ropensci/onboarding/issues/107

@arfon Could you please switch yourself to editor and proceed with accepting this submission?
