### 2017-08-14 #359 @arfon
> @arfon Could you please switch yourself to editor and proceed with accepting this submission?

Wilco.
