### 2017-08-18 #359 @AliciaSchep
Thanks @arfon and @karthik, and sorry for any confusion about the paper having gone through ropensci before, wasn't sure how this all worked!
